// server/src/enums/Eums.ts
export enum ErrorType {
    AUTH_ERROR = "AUTH_ERROR",
    DB_ERROR = "DB_ERROR",
    VALIDATION_ERROR = "VALIDATION_ERROR",
    NOT_FOUND_ERROR = "NOT_FOUND_ERROR",
    INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
    BAD_REQUEST_ERROR = "BAD_REQUEST_ERROR",
    CONFLICT_ERROR = "CONFLICT_ERROR",
    INVALID_TOKEN_ERROR = "INVALID_TOKEN_ERROR",
}